refactor(services): split interactions.rs into directory module
Closes the structural half of #90. Promotes the 627-line
crates/services/src/cell/interactions.rs (over the 500-line soft cap
in CLAUDE.md, with five obvious sibling functions) into an
interactions/ directory along the dispatch/dialog/vendor/trainer/loot
seams the issue specified.

  interactions/
  ├── mod.rs       # pub use facade
  ├── dispatch.rs  # handle_interact + handle_initial_response
  ├── dialog.rs    # send_dialog_display
  ├── vendor.rs    # send_store_open
  ├── trainer.rs   # send_trainer_open
  └── loot.rs      # send_loot_display + handle_loot_item

Public API unchanged: handle_interact, handle_initial_response,
send_dialog_display, and handle_loot_item are re-exported from mod.rs
so existing callers (cell::cell_methods::player::*, cell::content::
executor) keep their imports.

Internal helpers (send_store_open, send_trainer_open, send_loot_
display) are pub(super) — visible to dispatch.rs but not exposed
outside the module. Tests followed their owning function: arg-format
tests live with the wire-format helper, the four interact_* dispatch
tests live with handle_interact.

Cross-module path references (super::messages, super::space_manager,
super::abilities, super::content) became crate::cell::* now that the
file lives one level deeper.

No behavior change. cargo test -p cimmeria-services --lib
interactions: 7 passed.
